+++ activerecord/test/cases/helper.rb
@@ -205,5 +205,3 @@ module InTimeZone
       ActiveRecord::Base.time_zone_aware_attributes = old_tz
     end
 end
-
-require_relative "../../../tools/test_common"